xword-typeset: Allow circled squares
To place an open circle in any white square of the grid, change
its leading `|` character to `(`. This is the same thing that
"cwpuzzle.sty" version 1.11 calls `|[o]`, but I couldn't figure
out a nice way to make that syntax work.

There's no way to specify "circle this square" in the input file;
you just have to hack the LaTeX output by hand. Still, it's just
a one-character change per circle, so it's pretty straightforward.